Add BetterScreenshareAudio plugin
Implements a plugin that allows users to select audio sources for screensharing on Linux.

The plugin introduces a custom audio source selection menu within the Discord screenshare settings, allowing users to choose between the entire system audio or specific audio sources. It uses venmic to manage virtual microphones based on the selected audio sources.

Also includes an install script to automate the installation process for BetterDiscord and the plugin.